+++ tower-defense/tower-defense/src/com/git/tdgame/gameActor/projectile/HomingProjectile.java
@@ -66,7 +66,7 @@ public class HomingProjectile extends AbstractProjectile {
     	if(target instanceof Enemy)
     	{
     		Enemy e = (Enemy)target;
-    		e.takeDamage(damage);
+    		e.takeDamage((int)damage);
     		e.setProperty(slowAmount, slowDuration);
     		damagedEnemies.add(e);
     	}
@@ -91,7 +91,7 @@ public class HomingProjectile extends AbstractProjectile {
         	    	if(a instanceof Enemy && !damagedEnemies.contains(a, false))
         	    	{
         	    		Enemy e = (Enemy)a;
-        	    		e.takeDamage(damage);
+        	    		e.takeDamage((int)damage);
         	    		e.setProperty(slowAmount, slowDuration);
         	    		damagedEnemies.add(e);
         	    	}
